requirements.txt
Flask==3.0.2
gunicorn==21.2.0
psycopg2-binary
google-cloud-storage
